[PSym] Updates (#609)
* [PSym] Add swapping two machines in a value summary

* [PSym] track symmetrically-equivalent classes (instead of single class)

* [PSym] Adds first version of symmetry-aware exploration with symset merging

* [PSym] Adds first version of symmetry-aware symbolic exploration

* [PSym] Update symmetry tracker

Update symmetry tracker to exclude monitors during class merging

Adds tracking pending merges in symmetry tracker

Optimizes set summary contains() operation

Minor corrections

* [PSym] EMC: change learner to use double instead of BigDecimal

* [PSym] EMC: change state caching to hash instead of exact

* [PSym] refactor: separate explicit from symbolic (part 1)

* [PSym] refactor: rename scheduler

* [PSym] refactor: rename packages

* [PSym] refactor: remove verbose stats collection mode

* [PSym] refactor: separate out symbolic and explicit

+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/statistics/CoverageStats.java
@@ -4,9 +4,9 @@ import lombok.Getter;
 import lombok.Setter;
 import psym.runtime.logger.CoverageWriter;
 import psym.runtime.logger.StatWriter;
-import psym.runtime.scheduler.choiceorchestration.ChoiceLearningRewardMode;
-import psym.runtime.scheduler.choiceorchestration.ChoiceQTable;
-import psym.utils.GlobalData;
+import psym.runtime.scheduler.explicit.choiceorchestration.ChoiceLearningRewardMode;
+import psym.runtime.scheduler.explicit.choiceorchestration.ChoiceQTable;
+import psym.runtime.GlobalData;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
@@ -137,7 +137,7 @@ public class CoverageStats implements Serializable {
             for (int i = startDepth; i <= choiceDepth; i++) {
                 CoverageChoiceDepthStats stats = perChoiceDepthStats.get(i);
                 if (stats != null) {
-                    GlobalData.getChoiceLearningStats().rewardIteration(stats.getStateActions(), iterationCoverage, rewardMode);
+                    GlobalData.getChoiceLearningStats().rewardIteration(stats.getStateActions(), iterationCoverage.doubleValue(), rewardMode);
                 }
             }
         }
